requirements.txt
Django
djangorestframework
django-cors-headers
pandas
reportlab
PyQt5
requests
matplotlib